// front/src/app/modules/shared/model/LineaElement.ts
export interface LineaElement {
	id: number;
	codigo: string;
	nombre: string;
	cantidad: string;
	fecharegistro: string;
	linea: number;
	bodega: number;
	estado: string;
	desvincular: string;
	costo: number;
	costoventa: number;
	serial: string;
	costoventa2: number;
	costoventa3: number;
	iva: number;
	observacion: string;
	registro: string;
	registro2: string;
}